report/requirements.txt
altair==5.3.0
appnope==0.1.4
asn1crypto==1.5.1
astroid==3.1.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.103
botocore==1.34.103
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cramjam==2.8.3
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
fastparquet==2024.2.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydeck==0.9.1
Pygments==2.18.0
pylint==3.1.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
redshift-connector==2.1.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
scramp==1.4.5
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.34.0
tenacity==8.3.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
